Extract Java6Assertions superclass
